airflow/requirements.txt
pandas
apache-airflow
boto3
requests
numpy
python-dotenv
matplotlib
snowflake-connector-python
